Refatorando metodos e outros detalhes

+++ src/lib/outros.rs
@@ -1,5 +1,5 @@
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub struct Ponto{
-    pub x: i32,
-    pub y: i32
+    pub x: f64,
+    pub y: f64
 }
